Se implemento DoctorSpecialtyServiceTests

+++ src/main/java/com/thiagoporfirio/elvivo/domain/repositories/SpecialtyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SpecialtyRepository extends JpaRepository<SpecialtyEntity, Long>
+public interface SpecialtyRepository extends JpaRepository<SpecialtyEntity, Short>
 {
     public Optional<SpecialtyEntity> findByName(String name);
     public boolean existsByName(String name);
